infra/requirements.txt
aws-cdk-lib>=2.100.0
constructs>=10.0.0
pytest>=7.0.0
hypothesis>=6.0.0
